[PRODUCTION] Merge branch 'development' of https://github.com/smol-ai/talk into main (#29)
* add web scraping for submissions, add metadata migrations, fix type errors

* remove external resolver

* web scrape with cheerio and axios

* update artifact

* wrap req.json() in a try catch

* move scrape away from api route

* only run ci on actual pull requests

* 🛠️refactor(chat-layout) change import statement for Artifact type
🛠️refactor(chat-layout) add artifacts state and props to ChatLayout
🛠️refactor(chat-layout) pass artifacts to Sidebar component
⚙️chore(sidebar) modify Sidebar component to accept artifacts prop
⚙️chore(sidebar) add Discover tab to Tabs component
⚙️chore(sidebar) render list of artifacts in Discover tab
⚙️chore(sidebar) handle case when there are no artifacts to display
🚀feat(discover-page) create DiscoverPage component
🚀feat(discover-page) import necessary dependencies
🚀feat(discover-page) define DiscoverPageProps interface
🚀feat(discover-page) implement generateMetadata function
🚀feat(discover-page) implement DiscoverPage component
⚙️chore(layout) import getArtifacts function
⚙️chore(layout) call getArtifacts function and pass artifacts prop to ChatLayout
🛠️refactor(actions) add getArtifacts and getArtifact functions
🛠️refactor(artifact-item) create ArtifactItem component

* updating scrape to use scrapingbee

* change workflow to main and development

* switched function calls to next server actions (#23)

* switched function calls to next server actions

* Delete package-lock.json

* ignored notes.md for cleaner repository

* ignored .next from eslint and typescript

* Rate limiting, error message types, and new standardized chat and message types (#31)


* Add artifacts discover tab (#22)

* Update packages (#33)

* Implement PartyKit 🎈 (#39)

* add initial partykit implementation with ai

* update chat logic, move function calling to server side, add messages policies

* update global prompt (#40)

* 🚀feat(auth/session) add route handler for getting session
🛠️refactor(chat/scrape) rename createRouteHandlerClient to createServerComponentClient
⚙️chore(chat) remove commented out code

* 🚀feat(types) add Message type
⚙️chore(dependencies) upgrade @supabase/ssr to 0.0.7
⚙️chore(dependencies) upgrade partykit to 0.0.32
⚙️chore(dependencies) remove partykit 0.0.32 from package.json
⚙️chore(utils) rename getNextAuthSession to getAuthSession
⚙️chore(openai) add onFinalCallback to getChatCompletionResponse

* 🚀feat(server) add authentication and message handling
🛠️refactor(server) clean up code and remove unused imports
⚙️chore(server) update Supabase configuration

* migrate to openai

* set up auth

* WIP: Function Calling in PartyKit (#41)

* implementing function calling in partykit

* enable function calling in streaming output

---------


* backend fixes, update types

* updated types, remove useChat, return correct responses, update database

* In Progress: Function Calling Console (#43)

* update openai

* split out function calling

* formatting function response

* cleanup to function calling

* adding react-json-view to render function output

* wip: refactoring chat message component

* return parsed JSON content

* update role enum to reflect tool instead of function

* add status to functionCall type

* migrate to SmolTalkMessage type

* cleanup function output

---------


* Consolidate and document message types and their use (#45)

* consolidated and added documentation for message types

* implemented message logs in partykit workflow

* reafactoring smol talk chat hook + types

* More message refactoring + suppressing logs

* fix for LLM response race condition

* adding framer motion

* Fix console.log statements and add chat message
logs component


---------


* fix build types

* fix build

* deploy

* checkboxes for capabilities

* Add discover features (#44)

* 🛠️refactor(chat-list) fix type error
⚙️chore(discover-list) add new component
⚙️chore(empty-screen) update UI and add functionalities

* 🚀feat(discover-list) remove font size from article title
🛠️refactor(prompt-form) comment out ModelSelector component
⚙️chore(party/chat) update message insertion and update logic

* update artifacts schema, add tags, sources

* clean up ui

* 🛠️refactor(actions) update console.log to console

* 🚀feat(discover) add DiscoverArtifact component
⚙️chore(chat) add discover artifact to Chat component

* Refactor code to use recentArtifacts instead of
artifacts in EmptyScreen and Chat components

* 🚀feat(database) add new table 'scores'
🚀feat(database) add new table 'taxonomies'
🛠️refactor(database) rename table 'tags' to 'categorizations'
⚙️chore(database) modify table 'artifacts_tags' to 'artifacts_taxonomies'
⚙️chore(database) add timestamps to table 'artifacts_taxonomies'

* add database migrations, update types

* fix users table migration

---------


* Refactor identify function to include id parameter

* fix pnpm lock

* Add artifact ingest API endpoint (#50)

* add ingest api endpoint

* finished ingest to insert categories and topics

* fix

* delete types

* add types

* update node version to use latest

* fix ci?

* fix for ci cd test

---------

+++ app/actions.ts
@@ -1,7 +1,9 @@
 'use server'
 import 'server-only'
 
+import { Persona } from '@/constants/personas'
 import { Database } from '@/lib/db_types'
+import { Artifact, Message, SharedChat, type Chat, Taxonomy } from '@/lib/types'
 import {
   createServerActionClient,
   type User
@@ -9,32 +11,167 @@ import {
 import { revalidatePath } from 'next/cache'
 import { cookies } from 'next/headers'
 
-import { Persona } from '@/constants/personas'
-import { type Chat } from '@/lib/types'
-
-function nanoid() {
-  return Math.random().toString(36).slice(2) // random id up to 11 chars
-}
-
-export async function upsertChat(chat: Chat) {
+export async function insertChat(chat: Chat) {
   const cookieStore = cookies()
   const supabase = createServerActionClient<Database>({
     cookies: () => cookieStore
   })
 
-  const { error } = await supabase.from('chats').upsert({
-    id: chat.chat_id || nanoid(),
-    user_id: chat.userId,
-    payload: chat
-  })
+  const { data: insertedChat, error } = await supabase
+    .from('chats')
+    .insert({
+      ...(chat.id && { id: chat.id }),
+      user_id: chat.userId,
+      title: chat.title
+    })
+    .select()
+    .single()
+
   if (error) {
-    console.log('upsertChat error', error)
+    console.error('insertChat error', error)
     return {
       error: 'Unauthorized'
     }
-  } else {
+  }
+
+  const { error: messageError } = await supabase.from('messages').insert({
+    role: 'user',
+    user_id: chat.userId,
+    content: chat.messages[0].content,
+    chat_id: insertedChat.id
+  })
+
+  if (messageError) {
+    console.error('insertChat initial message error', error)
+    return {
+      error: 'Unauthorized'
+    }
+  }
+
+  return null
+}
+
+export async function getArtifacts() {
+  try {
+    const cookieStore = cookies()
+    const supabase = createServerActionClient<Database>({
+      cookies: () => cookieStore
+    })
+
+    const { data } = await supabase.from('artifacts').select(`
+        id,
+        title,
+        content,
+        created_at,
+        short_description,
+        taxonomies ( id, name, type )
+    `)
+    // .order('created_at', { ascending: false })
+    // .throwOnError()
+
+    const artifacts: Artifact[] =
+      data?.map(a => {
+        return {
+          id: a.id,
+          title: a.title,
+          content: a.content,
+          createdAt: a.created_at,
+          description: a.short_description,
+          category: a.taxonomies.find(t => t.type === 'category') as Taxonomy,
+          topics: a.taxonomies.filter(t => t.type === 'topic') as Taxonomy[],
+          tags: a.taxonomies.filter(t => t.type === 'tag') as Taxonomy[]
+        }
+      }) || []
+
+    return artifacts
+  } catch (error) {
+    console.error('get artifacts error', error)
+    return []
+  }
+}
+
+export async function getRecentArtifacts() {
+  try {
+    const cookieStore = cookies()
+    const supabase = createServerActionClient<Database>({
+      cookies: () => cookieStore
+    })
+
+    const date = new Date()
+    date.setDate(date.getDate() - 1)
+    const { data } = await supabase
+      .from('artifacts')
+      .select(
+        `
+        id,
+        title,
+        content,
+        created_at,
+        short_description,
+        taxonomies ( id, name, type )
+    `
+      )
+      .gte('created_at', date.toISOString())
+      .order('created_at', { ascending: false })
+      .throwOnError()
+
+    const artifacts: Artifact[] =
+      data?.map(a => {
+        return {
+          id: a.id,
+          title: a.title,
+          content: a.content,
+          createdAt: a.created_at,
+          description: a.short_description,
+          category: a.taxonomies.find(t => t.type === 'category') as Taxonomy,
+          topics: a.taxonomies.filter(t => t.type === 'topic') as Taxonomy[],
+          tags: a.taxonomies.filter(t => t.type === 'tag') as Taxonomy[]
+        }
+      }) || []
+
+    return artifacts
+  } catch (error) {
+    console.error('get artifacts error', error)
+    return []
+  }
+}
+
+export async function getArtifact(id: string) {
+  const cookieStore = cookies()
+  const supabase = createServerActionClient<Database>({
+    cookies: () => cookieStore
+  })
+  const { data } = await supabase
+    .from('artifacts')
+    .select(
+      `
+        id,
+        title,
+        content,
+        created_at,
+        short_description,
+        taxonomies ( id, name, type )
+    `
+    )
+    .eq('id', id)
+    .maybeSingle()
+
+  if (!data) {
     return null
   }
+
+  const artifact: Artifact = {
+    id: data.id,
+    title: data.title,
+    content: data.content,
+    createdAt: data.created_at,
+    description: data.short_description,
+    category: data.taxonomies.find(t => t.type === 'category') as Taxonomy,
+    topics: data.taxonomies.filter(t => t.type === 'topic') as Taxonomy[],
+    tags: data.taxonomies.filter(t => t.type === 'tag') as Taxonomy[]
+  }
+
+  return artifact
 }
 
 export async function getChats(userId?: string | null) {
@@ -48,12 +185,34 @@ export async function getChats(userId?: string | null) {
     })
     const { data } = await supabase
       .from('chats')
-      .select('payload')
-      .order('payload->createdAt', { ascending: false })
+      .select('id, title, created_at, updated_at, shared, messages ( id )')
+      .order('created_at', { ascending: false })
       .eq('user_id', userId)
+      .is('deleted_at', null)
+      .filter('messages.id', 'not.is', null)
       .throwOnError()
 
-    return (data?.map(entry => entry.payload) as Chat[]) ?? []
+    const chats: Chat[] =
+      data?.map(c => {
+        return {
+          id: c.id,
+          title: c.title,
+          path: `/chat/${c.id}`,
+          userId: userId,
+          messages: c.messages.map(m => {
+            return {
+              id: m.id
+            }
+          }) as Message[],
+          createdAt: new Date(c.created_at),
+          isShared: c.shared as boolean,
+          ...(c.shared && {
+            sharePath: `/share/${c.id}`
+          })
+        }
+      }) || []
+
+    return chats
   } catch (error) {
     return []
   }
@@ -64,13 +223,48 @@ export async function getChat(id: string) {
   const supabase = createServerActionClient<Database>({
     cookies: () => cookieStore
   })
-  const { data } = await supabase
+  const { data: chat } = await supabase
     .from('chats')
-    .select('payload')
+    .select(
+      'id, user_id, created_at, shared, title, messages (id, role, content, created_at)'
+    )
+    .order('created_at', { foreignTable: 'messages' })
+    .filter('messages.content', 'not.is', null)
     .eq('id', id)
+    .is('deleted_at', null)
     .maybeSingle()
 
-  return (data?.payload as Chat) ?? null
+  if (!chat) {
+    return null
+  }
+
+  const chatResponse = {
+    id: chat?.id,
+    title: chat?.title,
+    path: `/chat/${chat?.id}`,
+    userId: chat?.user_id,
+    messages: chat?.messages.map(m => {
+      return {
+        id: m.id,
+        role: m.role as
+          | 'function'
+          | 'user'
+          | 'assistant'
+          | 'system'
+          | 'tool'
+          | null,
+        content: m.content,
+        createdAt: new Date(m.created_at)
+      }
+    }) as Message[],
+    createdAt: new Date(chat.created_at),
+    isShared: chat.shared as boolean,
+    ...(chat.shared && {
+      sharePath: `/share/${chat.id}`
+    })
+  }
+
+  return chatResponse
 }
 
 export async function removeChat({ id, path }: { id: string; path: string }) {
@@ -79,7 +273,11 @@ export async function removeChat({ id, path }: { id: string; path: string }) {
     const supabase = createServerActionClient<Database>({
       cookies: () => cookieStore
     })
-    await supabase.from('chats').delete().eq('id', id).throwOnError()
+    await supabase
+      .from('chats')
+      .update({ deleted_at: new Date().toISOString() })
+      .eq('id', id)
+      .throwOnError()
 
     revalidatePath('/')
     return revalidatePath(path)
@@ -99,11 +297,17 @@ export async function clearChats() {
 
     const userId = (await supabase.auth.getUser())?.data?.user?.id
 
+    if (!userId) {
+      return {
+        error: 'Unauthorized'
+      }
+    }
+
     await supabase.from('chats').delete().eq('user_id', userId).throwOnError()
     revalidatePath('/')
     return revalidatePath('/')
   } catch (error) {
-    console.log('clear chats error', error)
+    console.error('clear chats error', error)
     return {
       error: 'Unauthorized'
     }
@@ -115,33 +319,106 @@ export async function getSharedChat(id: string) {
   const supabase = createServerActionClient<Database>({
     cookies: () => cookieStore
   })
-  const { data } = await supabase
+  const { data: chat } = await supabase
     .from('chats')
-    .select('payload')
+    .select(
+      'id, title, created_at, shared, messages(id, role, content, created_at)'
+    )
     .eq('id', id)
-    .not('payload->sharePath', 'is', null)
+    .is('shared', true)
     .maybeSingle()
 
-  return (data?.payload as Chat) ?? null
+  if (!chat) {
+    return null
+  }
+
+  const chatResponse: SharedChat = {
+    id: chat.id,
+    title: chat.title,
+    path: `/chat/${chat.id}`,
+    messages: chat.messages.map(m => {
+      return {
+        id: m.id,
+        role: m.role as 'user' | 'assistant' | 'system',
+        content: m.content as string,
+        createdAt: new Date(m.created_at)
+      }
+    }),
+    createdAt: new Date(chat.created_at)
+  }
+
+  return chatResponse
 }
 
 export async function shareChat(chat: Chat) {
-  const payload = {
-    ...chat,
-    sharePath: `/talk/share/${chat.id}`
+  try {
+    const cookieStore = cookies()
+    const supabase = createServerActionClient<Database>({
+      cookies: () => cookieStore
+    })
+
+    const { data } = await supabase
+      .from('chats')
+      .update({ shared: true })
+      .eq('id', chat.id)
+      .is('deleted_at', null)
+      .select('*, messages (id, role, content, created_at)')
+      .maybeSingle()
+      .throwOnError()
+
+    if (!data) {
+      throw new Error("Chat doesn't exist")
+    }
+
+    const chatResponse: Chat = {
+      id: data.id,
+      title: data.title,
+      path: `/chat/${data.id}`,
+      userId: data.user_id,
+      messages: data.messages.map(m => {
+        return {
+          id: m.id,
+          role: m.role as 'user' | 'assistant' | 'system',
+          content: m.content as string,
+          createdAt: new Date(m.created_at)
+        }
+      }),
+      createdAt: new Date(data.created_at),
+      isShared: data.shared as boolean,
+      ...(data.shared && {
+        sharePath: `/share/${data.id}`
+      })
+    }
+
+    return chatResponse
+  } catch (error) {
+    console.error(error)
+    return {
+      error: 'Unauthorized'
+    }
   }
+}
 
-  const cookieStore = cookies()
-  const supabase = createServerActionClient<Database>({
-    cookies: () => cookieStore
-  })
-  await supabase
-    .from('chats')
-    .update({ payload: payload as any })
-    .eq('id', chat.id)
-    .throwOnError()
+export async function getIsSubscribed(user: User) {
+  try {
+    const cookieStore = cookies()
+    const supabase = createServerActionClient<Database>({
+      cookies: () => cookieStore
+    })
 
-  return payload
+    const { data, error } = await supabase
+      .from('subscriptions')
+      .select('id')
+      .eq('user_id', user.id)
+      .filter('status', 'in', '("active", "trialing")')
+      .is('deleted_at', null)
+      .maybeSingle()
+  } catch (error) {
+    console.error('get is subscribed error', error)
+    return {
+      error: 'Unauthorized'
+    }
+  }
 }
 
 export async function getPersonas(user: User) {
@@ -152,8 +429,8 @@ export async function getPersonas(user: User) {
     })
 
     const { data, error } = await supabase
-      .from('prompts')
-      .select('id, prompt_name, prompt_body, emoji')
+      .from('personas')
+      .select('id, name, body, emoji')
       .order('created_at', { ascending: true })
       .eq('user_id', user.id)
       .is('deleted_at', null)
@@ -162,7 +439,7 @@ export async function getPersonas(user: User) {
 
     return personas
   } catch (error) {
-    console.log('get prompts error', error)
+    console.error('get personas error', error)
     return {
       error: 'Unauthorized'
     }
@@ -171,14 +448,18 @@ export async function getPersonas(user: User) {
 
 export async function getPersonaById(user: User, persona: Persona) {
   try {
+    if (!persona?.id) {
+      return null
+    }
+
     const cookieStore = cookies()
     const supabase = createServerActionClient<Database>({
       cookies: () => cookieStore
     })
 
     const { data, error } = await supabase
-      .from('prompts')
-      .select('id, prompt_name, prompt_body, emoji')
+      .from('personas')
+      .select('id, name, body, emoji')
       .eq('user_id', user.id)
       .eq('id', persona.id)
       .is('deleted_at', null)
@@ -188,7 +469,7 @@ export async function getPersonaById(user: User, persona: Persona) {
 
     return storedPersona
   } catch (error) {
-    console.log('get persona by ID error', error)
+    console.error('get persona by ID error', error)
     return {
       error: 'Unauthorized'
     }
@@ -205,10 +486,10 @@ export async function createOrUpdatePersona({
   try {
     // userData will update auth.users table
     const personaData = {
-      prompt_id: values.prompt_id,
-      prompt_name: values.prompt_name,
-      prompt_body: values.prompt_body,
-      prompt_emoji: values.prompt_emoji
+      id: values.id,
+      name: values.name,
+      body: values.body,
+      emoji: values.emoji
     }
 
     const cookieStore = cookies()
@@ -218,27 +499,27 @@ export async function createOrUpdatePersona({
 
     let result
 
-    if (personaData.prompt_id) {
+    if (personaData.id) {
       console.log('update persona', personaData)
       result = await supabase
-        .from('prompts')
+        .from('personas')
         .update({
           user_id: user.id,
-          prompt_name: personaData.prompt_name,
-          prompt_body: personaData.prompt_body,
-          emoji: personaData.prompt_emoji
+          name: personaData.name,
+          body: personaData.body,
+          emoji: personaData.emoji
         })
-        .eq('id', personaData.prompt_id)
+        .eq('id', personaData.id)
         .is('deleted_at', null)
         .select()
     } else {
       result = await supabase
-        .from('prompts')
+        .from('personas')
         .insert({
           user_id: user.id,
-          prompt_name: personaData.prompt_name,
-          prompt_body: personaData.prompt_body,
-          emoji: personaData.prompt_emoji
+          name: personaData.name,
+          body: personaData.body,
+          emoji: personaData.emoji
         })
         .eq('user_id', user.id)
         .select()
@@ -246,16 +527,16 @@ export async function createOrUpdatePersona({
     const { data: personaResponse, error } = result
 
     if (error) {
-      console.log('Error updating or adding persona:', error)
+      console.error('Error updating or adding persona:', error)
     }
 
     return {
       data: {
-        prompts: personaResponse
+        personas: personaResponse
       }
     }
   } catch (error) {
-    console.log('update persona error', error)
+    console.error('update persona error', error)
     return {
       error: 'Unauthorized'
     }
@@ -270,21 +551,21 @@ export async function removePersona({ id, user }: { id: string; user: User }) {
     })
 
     const { data: personaResponse, error } = await supabase
-      .from('prompts')
+      .from('personas')
       .update({ deleted_at: new Date().toISOString() })
       .eq('id', id)
 
     if (error) {
-      console.log('Error deleting persona:', error)
+      console.error('Error deleting persona:', error)
     }
 
     return {
       data: {
-        prompts: personaResponse
+        personas: personaResponse
       }
     }
   } catch (error) {
-    console.log('remove persona error', error)
+    console.error('remove persona error', error)
     return {
       error: 'Unauthorized'
     }
@@ -335,7 +616,7 @@ export async function updateUser({
       }
     }
   } catch (error) {
-    console.log('update user error', error)
+    console.error('update user error', error)
     return {
       error: 'Unauthorized'
     }
